// Harken-v2-release2025.08/Harken-v2-release2025.08/packages/backend/src/utils/enums/MessageEnum.ts
enum UserEnum {
	USER_NOT_FOUND = 'User not found',
	WRONG_PASSWORD = 'Wrong Password',
	LOGIN_SUCCESSFULLY = 'Logged in Successfully',
	USER_SAVE_SUCCESS = 'User Saved Successfully',
	ACCOUNT_DISABLED = 'Your account is disabled',
	INVALID_TOKEN = 'Token is invalid or expired',
	USER_INFO = 'User information',
	PASSWORD_NOT_UPDATE = 'There was a problem updating your password',
	PASSWORD_UPDATE = 'Your password has been updated. You may now login',
	TOKEN_REFRESHED = 'Your token has been refresh',
	PERMISSION_DENIED = 'Permission denied',
	USER_SAVE_FAIL = 'User Save Fail',
	INVALID_LOGIN = 'Login is invalid',
	NOT_APPROVED_BY_ADMIN = 'Your user has not been approved by an administrator yet',
	INVALID_ROLE = 'Invalid role',
	INVALID_ACCOUNT_ID = 'Invalid account id',
	ROLE_UPDATED_TO_ADMIN = 'Account details updated and The user role was updated to Account Administrator.',
	TOKEN_EXPIRED = 'Token is expired',
}
export enum MailEnum {
	MAIL_SEND = 'Check your inbox for a password reset email',
	FORGOT_PASSWORD_SUBJECT = 'Forgot password | Harken CRE',
	EMAIL_ALREADY_TAKEN = 'This email already taken',
	EMAIL_UPDATED_SUCCESSFULLY = 'User email updated successfully',
	EMAIL_NOT_TAKEN = 'Email not taken',
	SETTING_PASSWORD = 'Set your password | Harken CRE',
	ACCOUNT_MANAGER_CREATED = 'Account details updated and The account manager was created and instructions were sent to the defined email.',
	SEND_INVITE_SUBJECT = 'Invitation | Harken Bankable Data',
	SEND_INVITE = 'The invitations were sent!',
}